src/requirements.txt
aioredis==1.2.0
appnope==0.1.0
APScheduler==3.5.1
asgiref==3.2.1
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.8.1
Automat==0.7.0
awscli==1.16.92
awsebcli==3.14.9
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==3.1.0
blessed==1.15.0
boto==2.49.0
boto3==1.9.71
botocore==1.12.71
bs4==0.0.1
cached-property==1.5.1
cement==2.8.2
certifi==2018.4.16
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
colorama==0.3.9
constantly==15.1.0
cryptography==2.7
cycler==0.10.0
daphne==2.3.0
decorator==4.4.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.0.6
django-admin-interface==0.6.3
django-admin-list-filter-dropdown==1.0.1
django-admin-sortable==2.1.7
django-annoying==0.10.4
django-ckeditor==5.6.1
django-colorfield==0.1.15
django-cors-headers==2.3.0
django-crispy-forms==1.7.2
django-email-gateway==0.1.2
django-filter==2.0.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==1.1.0
django-ses==0.8.7
django-ses-gateway==0.1.1
django-softdelete-it==0.1
django-summernote==0.8.8.8
django-webpack-loader==0.6.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
docker==3.7.0
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
entrypoints==0.3
future==0.17.1
googlemaps==3.0.2
graphviz==0.9
gunicorn==19.9.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.7
incremental==17.5.0
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
jedi==0.14.1
Jinja2==2.10.1
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.0.1
lxml==4.2.5
MarkupSafe==1.1.1
matplotlib==2.2.3
mistune==0.8.4
msgpack==0.6.1
mysqlclient==1.3.13
nbconvert==5.5.0
nbformat==4.4.0
notebook==6.0.0
numpy==1.14.5
olefile==0.45.1
pandas==0.23.3
pandocfilters==1.4.2
parso==0.5.1
pathspec==0.5.5
pexpect==4.7.0
pickleshare==0.7.5
Pillow==5.2.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psycopg2==2.7.5
ptyprocess==0.6.0
pyasn1==0.4.5
pycparser==2.19
pycrypto==2.6.1
Pygments==2.4.2
PyHamcrest==1.9.0
PyJWT==1.7.1
pyparsing==2.2.0
python-dateutil==2.7.5
pytz==2018.5
PyYAML==3.13
pyzmq==18.0.2
qtconsole==4.5.1
razorpay==1.1.1
requests==2.19.1
rsa==3.4.2
s3transfer==0.1.13
scikit-learn==0.19.2
scipy==1.1.0
seaborn==0.9.0
selenium==3.141.0
semantic-version==2.5.0
Send2Trash==1.5.0
six==1.12.0
sqlparse==0.3.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
texttable==0.9.1
tornado==6.0.3
traitlets==4.3.2
Twisted==19.7.0
txaio==18.8.1
tzlocal==1.5.1
urllib3==1.23
virtualenv==16.0.0
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.54.0
widgetsnbextension==3.5.0
xlrd==1.2.0
zope.interface==4.6.0
